feat: Upgrade Library Versions and Python Version (#187)
- moving to Python 3.10
- Upgrade all library versions in requirements.txt and setup.py
- Remove DM-Haiku cause it is now deprecated.
- all networks now are based on a single MLP class
- jax.tree_map has been replaced with jax.tree_util.tree_map everywhere to avoid Deprecation Warnings.
- types -> custom_types
- add extra_require for jax[cuda12]
- fix all notebooks and update typing extensions (for running notebooks)
- fix dependabot security issues
- added instructions for pip install qdax[cuda12] in README
- fix observation space in jumanji test script

+++ tests/core_test/mome_test.py
@@ -14,7 +14,7 @@ from qdax.core.emitters.mutation_operators import (
 )
 from qdax.core.emitters.standard_emitters import MixingEmitter
 from qdax.core.mome import MOME
-from qdax.types import Descriptor, ExtraScores, Fitness, RNGKey
+from qdax.custom_types import Descriptor, ExtraScores, Fitness, RNGKey
 from qdax.utils.metrics import default_moqd_metrics
 
 
@@ -36,10 +36,10 @@ def test_mome(num_descriptors: int) -> None:
     crossover_percentage = 1.0
     batch_size = 80
     lag = 2.2
-    base_lag = 0
+    base_lag = 0.0
 
     def rastrigin_scorer(
-        genotypes: jnp.ndarray, base_lag: int, lag: int
+        genotypes: jnp.ndarray, base_lag: float, lag: float
     ) -> Tuple[jnp.ndarray, jnp.ndarray]:
         """
         Rastrigin Scorer with first two dimensions as descriptors
@@ -131,7 +131,11 @@ def test_mome(num_descriptors: int) -> None:
     )
 
     # Run the algorithm
-    (repertoire, emitter_state, random_key,), metrics = jax.lax.scan(
+    (
+        repertoire,
+        emitter_state,
+        random_key,
+    ), metrics = jax.lax.scan(
         mome.scan_update,
         (repertoire, emitter_state, random_key),
         (),
